// server/src/user/dto/update-user.dto.d.ts
export declare class UpdateUserDto {
    firstName?: string;
    lastName?: string;
    phone?: string;
    tag?: string;
}
export declare class UpdateKycDto {
    kycStatus: string;
}
export declare class UpdatePinDto {
    pin: string;
}
export declare class UpdateBiometricDto {
    biometricEnabled: boolean;
}